+++ google/api/iam/methods.ts
@@ -1,8 +1,10 @@
 import { methods as projectMethods } from './projects';
+import { methods as serviceAccountMethods } from './service-account';
 
 /**
  * IAM API Methods
  */
 export default {
   projects: projectMethods,
+  serviceAccount: serviceAccountMethods,
 };