+++ google/index.ts
@@ -1,7 +1,7 @@
 import { resolveToken, stringizeScope } from './helpers';
 
 export { api as oauthAuthenticationApis } from './helpers';
-export { api as oauthRouteApis } from './API';
+export { api as oauthRouteApis } from './routes';
 
 export { default as normalAccountHandler } from './handlers/nac';
 export { default as serviceAccountHandler } from './handlers/sac';